% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_release_date.R
\name{abs_current_release}
\alias{abs_current_release}
\title{Return the date that the most current
version of an ABS Time Series was released}
\usage{
abs_current_release(cat_no)
}
\arguments{
\item{cat_no}{string. Include the .0}
}
\description{
Return the date that the most current
version of an ABS Time Series was released
}
