% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_hoursworked}
\alias{abs_hoursworked}
\title{Create a plot for the monthly hours worked in all jobs from the ABS labour force series.}
\usage{
abs_hoursworked(
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{states}{Australia or any state/territory in Australia.}

\item{years}{The number of years overwhich to present the data. Available options are 1,3,5,10, and 20 years.}

\item{compare_aus}{If TRUE (the default) adds the Australian data to the plot. Values are indexed to January of the current year minus "years"}

\item{ages}{default "Total (age)"}

\item{genders}{default "Persons"}

\item{series_types}{default "Trend"}
}
\value{
ggplot2 object
}
\description{
Create a plot for the monthly hours worked in all jobs from the ABS labour force series.
}
