% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plot.R
\name{abs_plot}
\alias{abs_plot}
\alias{abs_plot_wpi}
\alias{abs_plot_labour_force}
\title{Generate a time-series plot for ABS Time Series Data.}
\usage{
abs_plot(data = NULL, filter_with, type, years = 2015, ...)

abs_plot_wpi(
  data,
  filter_with,
  years,
  index_to = "2020-03-01",
  compare_aus = FALSE
)

abs_plot_labour_force(
  data,
  filter_with,
  years,
  compare_aus = TRUE,
  markdown = FALSE,
  facet = NULL,
  plotly = FALSE,
  void = FALSE
)
}
\arguments{
\item{data}{ABS Time Series data}

\item{filter_with}{named list specifying what should be plot.}

\item{type}{One of "wpi" for Wage Price Index plots or "labour_force" for Labour Force plots.}

\item{years}{How many years of data to plot, defaults to years.}

\item{...}{other arguments passed to various functions depending on type.}

\item{index_to}{Date as "yyyy-mm-dd" to re-index data to, defaults to "2020-03-01"}

\item{compare_aus}{Whether to automatically add Australian data to plot, defaults to FALSE}

\item{markdown}{Whether to use markdown formatting of titles and subtitles. NYI.}

\item{facet}{Whether to facet the plot to include additional dimensions. NYI.}

\item{plotly}{Whether to use \code{ggplotly} to create a plotly interactive plot. NYI.}

\item{void}{Whether to strip all theme elements from the plot.}
}
\value{
A ggplot2 time-series plot or a plotly time-series plot if plotly = TRUE
}
\description{
\code{abs_plot()} can automatically generate appropriate plots for ABS
Time Series indicators for both static display in documents, or RMarkdown,
as well as interactive plots through plotly.

\code{\link[=abs_plot_wpi]{abs_plot_wpi()}}, \code{\link[=abs_plot_labour_force]{abs_plot_labour_force()}} can be called directly and
create plots for WPI and Labour Force data respectively.
}
\examples{
\dontrun{
#Plot employed total for Australia and South Australia
abs_plot(filter_with = list(indicator = "Employed total", state = c("South Australia", "Australia")), type = "labour_force")
}

}
\seealso{
\itemize{
\item \link{abs_plot_wpi} for WPI plots.
\item \link{abs_plot_labour_force} for Labour Force plots.
}
}
\concept{abs plot functions}
