% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plotly.R
\name{abs_plotly}
\alias{abs_plotly}
\title{Draw interactive abs plots}
\usage{
abs_plotly(
  indicator,
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_type = "Seasonally Adjusted"
)
}
\arguments{
\item{indicator}{string. One of:

\itemize{
\item{"Employment"}
\item{"Unemployment"}
\item{"Participation rate"}
}}

\item{series_type}{}
}
\value{

}
\description{
Draw interactive abs plots
}
