% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_release_date.R
\name{abs_release_date}
\alias{abs_release_date}
\title{Next Release date of an ABS Time Series}
\usage{
abs_release_date(cat_no)
}
\arguments{
\item{cat_no}{string. include the .0}
}
\description{
Next Release date of an ABS Time Series
}
