% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_underemployment_rate}
\alias{abs_underemployment_rate}
\title{Title}
\usage{
abs_underemployment_rate(
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{series_types}{}
}
\value{

}
\description{
Title
}
