% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_underutilisation_rate}
\alias{abs_underutilisation_rate}
\title{Draw historical underutilisation rates}
\usage{
abs_underutilisation_rate(
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{series_types}{}
}
\value{

}
\description{
Draw historical underutilisation rates
}
