% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_unemployment}
\alias{abs_unemployment}
\title{Unemployment (level) Time Series}
\usage{
abs_unemployment(
  states,
  years,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{dbf}{}
}
\value{

}
\description{
Unemployment (level) Time Series
}
