% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_unemployment_rate}
\alias{abs_unemployment_rate}
\title{ABS Unemployment Rate Time Series}
\usage{
abs_unemployment_rate(
  states,
  years,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{dbf}{}
}
\value{

}
\description{
ABS Unemployment Rate Time Series
}
