% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_palettes.R
\name{aiti_cols}
\alias{aiti_cols}
\title{List available colours}
\usage{
aiti_cols(colour = NULL)
}
\arguments{
\item{colour}{A character of the name of the colour. If NULL, returns all available colours.}
}
\value{
A named character vector of the defined colours in the package.
}
\description{
List available colours
}
\examples{
aiti_cols()
}
