% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_palettes.R
\name{aiti_pal}
\alias{aiti_pal}
\title{Generate AITI colours from a palette}
\usage{
aiti_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette}

\item{reverse}{TRUE to reverse the order of the colours}

\item{...}{Other arguments passed to `colorRampPalette()`}
}
\description{
Generate AITI colours from a palette
}
