% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_amwu_colours.R
\docType{data}
\name{amwu_darkblue6}
\alias{amwu_darkblue6}
\title{amwu_darkblue6}
\format{
An object of class \code{character} of length 1.
}
\usage{
amwu_darkblue6
}
\description{
amwu_darkblue6
}
\keyword{datasets}
