% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_amwu_colours.R
\docType{data}
\name{amwu_yellow1}
\alias{amwu_yellow1}
\title{amwu_yellow1}
\format{
An object of class \code{character} of length 1.
}
\usage{
amwu_yellow1
}
\description{
amwu_yellow1
}
\keyword{datasets}
