% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_amwu_colours.R
\docType{data}
\name{amwu_yellow2}
\alias{amwu_yellow2}
\title{amwu_yellow2}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
amwu_yellow2

amwu_yellow2
}
\description{
amwu_yellow2

amwu_yellow2
}
\keyword{datasets}
