% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format.R
\name{as_comma_group}
\alias{as_comma_group}
\title{Format a group of values of different magnitudes with comma separation}
\usage{
as_comma_group(.data, group = NULL, value = NULL, suffix = NULL, digits = 0)
}
\arguments{
\item{.data}{A dataframe which contains the group of values}

\item{group}{Character. A column in the supplied dataframe specifying which group the values belong.
If NULL, row numbers will be used. See the example for more details.}

\item{value}{Numeric. A column in the supplied dataframe which contains the values to be comma separated.}

\item{suffix}{Character. An optional character to print at the end of each value, i.e. "persons".}

\item{digits}{Numeric. The number of digits}
}
\value{
A character vector of the same length as the input dataframe.
}
\description{
Format a group of values of different magnitudes with comma separation
}
