% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format.R
\name{as_percent}
\alias{as_percent}
\title{Format a value as a pretty string, as a percentage}
\usage{
as_percent(string, scale = 1, ...)
}
\arguments{
\item{string}{Value to print}

\item{scale}{Scale the value by a multiple}

\item{...}{Other arguments passed to formatC}
}
\value{
character
}
\description{
Format a value as a pretty string, as a percentage
}
\examples{
as_percent(50) = "50\%"
}
