% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format.R
\name{as_percentage_point}
\alias{as_percentage_point}
\title{Format a value as a pretty string, with the percentage points suffix}
\usage{
as_percentage_point(string, scale = 1, digits = 1)
}
\arguments{
\item{string}{the value to print}

\item{scale}{A value to multiply the number before converting to a string. Default of 1 assumes the value
to be printed has already been multipled by 100.}

\item{digits}{Number of digits to print. Defaults to 1}
}
\value{
character
}
\description{
Format a value as a pretty string, with the percentage points suffix
}
\examples{
as_percentage_point(1.5)
}
