% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{average_over}
\alias{average_over}
\title{Title}
\usage{
average_over(data = .data, filter_with = filter_list, between)
}
\arguments{
\item{between}{}
}
\value{

}
\description{
Title
}
\examples{

}
