% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalab_clearance.R
\name{clear_output}
\alias{clear_output}
\title{Apply dominance check to ABS unit record data. This function is not exported}
\usage{
clear_output(data, summary_var, by)
}
\arguments{
\item{data}{A data frame}

\item{summary_var}{The variables needing output clearance}

\item{by}{group by}
}
\value{
data frame
}
\description{
Apply dominance check to ABS unit record data. This function is not exported
}
