% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price-utils.R
\name{cumulative_cpi}
\alias{cumulative_cpi}
\title{Create Cumulative CPI Data Set}
\usage{
cumulative_cpi(data, region, since, include_housing = F)
}
\arguments{
\item{data}{CPI or CLI data from \code{read_absdata}}

\item{region}{The capital city or Australia}

\item{since}{Date ("\%Y-\%m-\%d") to benchmark}

\item{include_housing}{whether to include mortgage repayments (requires data = CLI)}
}
\value{
dataframe
}
\description{
Create Cumulative CPI Data Set
}
