% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price-utils.R
\name{cumulative_wpi}
\alias{cumulative_wpi}
\title{Create Cumulative WPI data set}
\usage{
cumulative_wpi(
  data,
  state,
  indicator = "Total hourly rates of pay excluding bonuses",
  sector = "Private and Public",
  industry = "All industries",
  since
)
}
\arguments{
\item{data}{WPI data from \code{read_absdata}}

\item{state}{The state, or Australia}

\item{indicator}{Indicator}

\item{sector}{Sector}

\item{industry}{Industry}

\item{since}{Date ("\%Y-\%m-\%d") to benchmark}
}
\description{
Create Cumulative WPI data set
}
