% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{current}
\alias{current}
\title{Return the most recent value for an indicator}
\usage{
current(data = .data, filter_with = filter_list, print = TRUE)
}
\arguments{
\item{filter_with}{}
}
\value{
num
}
\description{
Return the most recent value for an indicator
}
