% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{employment_growth}
\alias{employment_growth}
\title{Convenience function to draw a bar chart of employment growth, by different types}
\usage{
employment_growth(states, series_types, since = 2010)
}
\arguments{
\item{states}{Which state to plot}

\item{series_types}{Which series type to plot}

\item{since}{Growth of employment since a given year}
}
\value{
ggplot2 object
}
\description{
Convenience function to draw a bar chart of employment growth, by different types
}
