% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_with.R
\name{filter_with}
\alias{filter_with}
\title{Filter a dataframe with semi-consistent column names (ie some series do not have a "gender") and semi-consistent
variables within columns (ie some series do not have a trend indicator for some states).
This function is not exported}
\usage{
filter_with(data = NULL, filter_with = NULL)
}
\description{
Filter a dataframe with semi-consistent column names (ie some series do not have a "gender") and semi-consistent
variables within columns (ie some series do not have a trend indicator for some states).
This function is not exported
}
