% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fof_colours}
\alias{fof_colours}
\title{Factory of the Future Colours}
\format{
An object of class \code{character} of length 11.
}
\usage{
fof_colours
}
\description{
Factory of the Future Colours
}
\keyword{datasets}
