% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fof_palettes.R
\name{fof_pal}
\alias{fof_pal}
\title{Generate Factory of the Future colours from a pre-defined palette}
\usage{
fof_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette. See \code{palette_names()} for a list of available palettes.}

\item{reverse}{Whether the order of the colours should be reversed (default FALSE)}

\item{...}{Other arguments passed to \code{colorRampPalette()}}
}
\value{
A colour palette
}
\description{
Generate Factory of the Future colours from a pre-defined palette
}
