% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fof_palettes}
\alias{fof_palettes}
\title{Factory of the Future Colour Palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
fof_palettes
}
\description{
Factory of the Future Colour Palettes
}
\keyword{datasets}
