% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abs_datacube.R
\name{get_abs_datacube}
\alias{get_abs_datacube}
\title{Title}
\usage{
get_abs_datacube(series, cube, path = "data-raw", latest = TRUE)
}
\arguments{
\item{latest}{}
}
\value{

}
\description{
Title
}
\examples{

}
