% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Title}
\usage{
growth(data = NULL, filter_with = filter_list, ym = "year")
}
\arguments{
\item{data}{A dataframe of tidyed ABS data}

\item{filter_with}{A list of which variables to filter the dataframe by. Must be a list. Can specify indicator, gender, state, age, and series_type. Indicator must be specified}

\item{year_since}{Baseline growth year}
}
\value{
dataframe
}
\description{
Title
}
\examples{
growth(underutilisation, filter_with = list(indicator = "Unemployment rate"))

}
