% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobkeeper_plots.R
\name{jobkeeper_plots}
\alias{jobkeeper_plots}
\title{Jobkeeper plots}
\usage{
jobkeeper_plots(
  data = NULL,
  indicator,
  states,
  years,
  compare_aus = TRUE,
  plotly = FALSE,
  void = FALSE
)
}
\arguments{
\item{void}{}
}
\value{

}
\description{
Jobkeeper plots
}
