% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labour_force}
\alias{labour_force}
\title{Births}
\format{A data frame:
\describe{
\item{year}{Year}
\item{births}{Number of live births, rounded to nearest 1000}}}
\source{
\url{www.abs.gov.au}
}
\usage{
labour_force
}
\description{
Live births data from census.gov.
}
\keyword{datasets}
