% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{last}
\alias{last}
\title{Reporting functions}
\usage{
last(data = .data, filter_with = filter_list, ym = "year")
}
\arguments{
\item{filter_with}{}
}
\value{

}
\description{
Reporting functions
}
\examples{

}
