% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{last_value}
\alias{last_value}
\title{Find the value for an indicator for this month last year, or the previous month.}
\usage{
last_value(data = NULL, filter_with = filter_list, ym = "year", print = TRUE)
}
\arguments{
\item{data}{a dataframe of cleaned ABS Time Series data returned from readabs}

\item{filter_with}{a list of variables to filter the dataframe on. Valid variables include
gender, age, indicator, and series type.}

\item{ym}{'year' to return the value this time last year, 'month' to return the value for this time last month}

\item{print}{logical to pass the numeric on to as_comma or as_percent for printing. Defaults to TRUE}
}
\value{
numeric if print == FALSE, character if print == TRUE
}
\description{
Find the value for an indicator for this month last year, or the previous month.
}
\examples{
last_value()

}
