% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manufacturing_alliance}
\alias{manufacturing_alliance}
\title{Manufacturing Alliance Definitions}
\format{
\subsection{\code{manufacturing_alliance}}{

A data frame with 166 rows and 9 columns:
\describe{
\item{anzsic_division_code}{AZSIC 2006 Division Code}
\item{anzsic_division}{ANZSIC 2006 Division Name}
\item{anzsic_subdivision_code}{ANZSIC 2006 Subdivision Code}
\item{anzsic_subdivision}{ANZSIC 2006 Subdivision Name}
\item{anzsic_group_code}{ANZISC 2006 Group Code}
\item{anzsic_group}{ANZSIC 2006 Group Name}
\item{anzsic_class_code}{ANZSIC 2006 Class Code}
\item{anzsic_class}{ANZSIC 2006 Class Name}
\item{sub_sector}{Manufacturing Alliance Sub-Sector}
}
@source \url{https://manufacturingalliance.org.au/wp-content/uploads/2024/07/Appendix-A-Defining-the-Manufacturing-Alliances-Workforce.pdf}
}
}
\usage{
manufacturing_alliance
}
\description{
A subset of ANZSIC2006 which captures the Manufacturing Alliance
definitions of manufacturing sub-sectors.
}
\keyword{datasets}
