% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_palettes.R
\name{nrc_cols}
\alias{nrc_cols}
\title{List available colours}
\usage{
nrc_cols(colour = NULL)
}
\arguments{
\item{colour}{A character of the name of the colour. If NULL, returns all available colours.}
}
\value{
A named character vector of the defined colours in the package.
}
\description{
List available colours
}
\examples{
nrc_cols()
}
