% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_palettes.R
\name{nrc_pal}
\alias{nrc_pal}
\title{Generate Australian Manufacturing Workers Union colours from a pre-defined palette}
\usage{
nrc_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette. See \code{palette_names()} for a list of available palettes.}

\item{reverse}{Whether the order of the colours should be reversed (default FALSE)}

\item{...}{Other arguments passed to \code{colorRampPalette()}}
}
\value{
A colour palette
}
\description{
Generate Australian Manufacturing Workers Union colours from a pre-defined palette
}
\examples{
To generate a sequence of n colours from the 'main' palette:
n <- 2
nrc_pal("main")(n)
}
