% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_palettes.R
\name{palette_names}
\alias{palette_names}
\title{List available palettes}
\usage{
palette_names()
}
\value{
character vector of palette names
}
\description{
List available palettes
}
\examples{
palette_names()
}
