% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dates.R
\name{plot_date_range}
\alias{plot_date_range}
\title{Create Date Range for Plots}
\usage{
plot_date_range(plot_data)
}
\arguments{
\item{plot_data}{The data}
}
\value{
a vector of dates
}
\description{
Ensures the last period of data is shown on the plot.
}
\examples{
library(ggplot2)

ggplot(economics, aes(x = date, y = uempmed)) +
geom_line() +
scale_x_date(breaks = plot_date_range(economics))

}
