% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_employed_total}
\alias{plot_employed_total}
\alias{plot_employed_full_time}
\alias{plot_employed_part_time}
\alias{plot_unemployed_total}
\alias{plot_unemployed_looked_for_full_time_work}
\alias{plot_unemployed_looked_for_part_time_work}
\alias{plot_underemployed_total}
\alias{plot_underutilised_total}
\alias{plot_hours_worked}
\alias{plot_hours_worked_full_time}
\alias{plot_hours_worked_part_time}
\alias{plot_not_in_the_labour_force}
\alias{plot_working_population}
\alias{plot_employment_population_ratio}
\alias{plot_unemployment_rate}
\alias{plot_unemployment_rate_looked_for_full_time_work}
\alias{plot_unemployment_rate_looked_for_part_time_work}
\alias{plot_participation_rate}
\alias{plot_underemployment_ratio}
\alias{plot_underemployment_rate}
\alias{plot_underutilisation_rate}
\alias{plot_sahm_recession_indicator}
\alias{plot_jobkeeper_applications}
\title{Convenience functions for creating charts for ABS and other}
\usage{
plot_employed_total(states, ...)

plot_employed_full_time(states, ...)

plot_employed_part_time(states, ...)

plot_unemployed_total(states, ...)

plot_unemployed_looked_for_full_time_work(states, ...)

plot_unemployed_looked_for_part_time_work(states, ...)

plot_underemployed_total(states, ...)

plot_underutilised_total(states, ...)

plot_hours_worked(states, ...)

plot_hours_worked_full_time(states, ...)

plot_hours_worked_part_time(states, ...)

plot_not_in_the_labour_force(states, ...)

plot_working_population(states, ...)

plot_employment_population_ratio(states, ...)

plot_unemployment_rate(states, ...)

plot_unemployment_rate_looked_for_full_time_work(states, ...)

plot_unemployment_rate_looked_for_part_time_work(states, ...)

plot_participation_rate(states, ...)

plot_underemployment_ratio(states, ...)

plot_underemployment_rate(states, ...)

plot_underutilisation_rate(states, ...)

plot_sahm_recession_indicator(states, ...)

plot_jobkeeper_applications(states)
}
\arguments{
\item{states}{The states to include in the plot.}

\item{...}{Other options passed to abs_plot}
}
\description{
These functions are wrap-around functions for abs_plot() for most of the
indicators you should ever need to plot. For more information about what options you can specify see
\code{\link{abs_plot}}
}
