% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price-utils.R
\name{plot_real_wages}
\alias{plot_real_wages}
\title{Plot Real Wages Over Time}
\usage{
plot_real_wages(
  cpi_data,
  wpi_data,
  forecast = list(years = 1, wage = 10, inflation = 3),
  ...
)
}
\arguments{
\item{cpi_data}{data from \code{read_absdata}}

\item{wpi_data}{data from \code{read_absdata}}

\item{forecast}{(optional) forecast of inflation and wages \code{list(years, wage, inflation)}}

\item{...}{other arguments passed on to functions}
}
\value{
ggplot2 graph
}
\description{
Plot Real Wages Over Time
}
\examples{
\dontrun{
plot_real_wages(read_absdata("cpi_quarterly_group"),
                read_absdata("wpi_quarterly"),
                forecast = NULL,
                region = "Australia",
                state = "Australia",
                since = "2020-03-01",
                include_housing = FALSE)
                }
}
