% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_absdata.R
\name{read_absdata}
\alias{read_absdata}
\title{Read tidied ABS data from the aitidata GitHub repository.}
\usage{
read_absdata(
  name = NULL,
  refresh = FALSE,
  export_dir = tempdir(),
  .validate_name = TRUE
)
}
\arguments{
\item{name}{The name, as a string, of the dataset to download.}

\item{refresh}{FALSE (the default) will load a cached version of the data if it has already been downloaded.
Specify TRUE to force the data to be downloaded again.}

\item{export_dir}{The directory to save downloaded data in. Defaults to a temporary directory.}

\item{.validate_name}{NYI}
}
\value{
data
}
\description{
Read tidied ABS data from the aitidata GitHub repository.
}
