% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportabs-package.R
\docType{package}
\name{reportabs-package}
\alias{reportabs}
\alias{reportabs-package}
\title{reportabs: Makes Reporting ABS Time Series Data Simpler}
\description{
Makes reporting ABS TimeSeries data simpler. It provides convenience functions for plotting labour market indicators, and for reporting of changes in indicators over time. Links with the aitidata package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://f-fof.github.io/reportabs/}
}

}
\author{
\strong{Maintainer}: Hamish Gamble \email{hamish.gamble@flinders.edu.au}

}
\keyword{internal}
