% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scale.R
\name{scale_colour_aiti}
\alias{scale_colour_aiti}
\title{AITI Colour Scale}
\usage{
scale_colour_aiti(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Name of colour palette}

\item{discrete}{TRUE for a discrete scale}

\item{reverse}{TRUE to reverse the order of the colours}
}
\value{
ggplot2 colour scale
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{scale_colour_aiti()} has been renamed \code{scale_colour_fof()}
}
