% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scale.R
\name{scale_fill_fof}
\alias{scale_fill_fof}
\title{Fill scale for Factory of the Future colours}
\usage{
scale_fill_fof(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Name of the colour palette (default "main")}

\item{discrete}{Whether the colours should be discrete or continuous (default TRUE)}

\item{reverse}{Whether the order of the colours should be reversed (default FALSE)}
}
\value{
ggplot2 fill scale
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{scale_fill_fof()} has been renamed \code{scale_fill_nrc()}
}
