% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_living.R
\name{table_cli}
\alias{table_cli}
\title{Change in Living Costs}
\usage{
table_cli(wpi_data, cli_data, since)
}
\arguments{
\item{wpi_data}{Wage Price Index (wpi_quarterly) data from \code{read_absdata}}

\item{cli_data}{Living Cost Index data (cli) from \code{read_absdata}}

\item{since}{The date ("\%Y-\%m-\%d) for comparison}
}
\value{
A gt table
}
\description{
A \code{gt} table of the change in living costs, by category, between the most recent
available data and the date specified in \code{since}. Includes the change in Manufacturing
industry wages over the same period for comparison.
}
\examples{
\dontrun{
wpi_data <- read_absdata("wpi_quarterly", export_dir = "data")
cli_data <- read_absdata("cli", export_dir = "data")
table_cli(wpi_data, cli_data, "2020-03-01")
}
}
