% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aiti.R
\name{theme_aiti}
\alias{theme_aiti}
\title{AITI Theme}
\usage{
theme_aiti(
  base_size = 12,
  colour = "Soft Black",
  legend = "none",
  markdown = FALSE,
  flipped = FALSE
)
}
\arguments{
\item{base_size}{The base size of text elements of the plot.}

\item{colour}{The background colour of the plot (at 20\% transparency).}

\item{legend}{Where to position the legend. Default "none" will not show a legend.}

\item{markdown}{Use markdown for text elements.}

\item{flipped}{TRUE to flip the y-axis guide lines to show on the x-axis instead.}
}
\value{
ggplot2 theme
}
\description{
Theme for plots in AITI publications, reports, Shiny Apps.
}
