% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_fof}
\alias{theme_fof}
\title{ggplot2 theme for Factory of the Future}
\usage{
theme_fof(
  base_size = 12,
  colour = "Soft Black",
  legend = "none",
  markdown = FALSE,
  flipped = FALSE,
  legacy = FALSE
)
}
\arguments{
\item{base_size}{The base size of text elements of the plot. (default 12)}

\item{colour}{The background colour of the plot. Background colours are applied at 80\% transparency. (default 'Soft Black')}

\item{legend}{The position of the legend. (default "none")}

\item{markdown}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Whether to use markdown formatting for plot titles (default FALSE)}

\item{flipped}{Whether to flip the y-axis guide lines to show on the x-axis instead (default FALSE)}

\item{legacy}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Whether to use legacy (AITI) fonts. This option will be removed sometime in the future.}
}
\value{
a ggplo2 theme
}
\description{
Theme for plots in publications, reports, Shiny apps, etc. Anywhere ggplot2 is used, you can use this theme!
}
