% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_nrc}
\alias{theme_nrc}
\title{ggplot2 theme for Australian Manufacturing Workers Union}
\usage{
theme_nrc(
  base_size = 12,
  legend = "none",
  ink = midnight,
  paper = white,
  flipped = FALSE
)
}
\arguments{
\item{base_size}{The base size of text elements of the plot. (default 12)}

\item{legend}{The position of the legend. (default "none")}

\item{ink}{Ink (foreground) colour. (default midnight)}

\item{paper}{Paper (background) colour. (default white)}

\item{flipped}{Whether to flip the y-axis guide lines to show on the x-axis instead (default FALSE)}
}
\value{
a ggplot2 theme
}
\description{
Theme for plots in publications, reports, Shiny apps, etc. Anywhere ggplot2 is used, you can use this theme!
}
\examples{
\dontrun{
library(ggplot2)
df <- data.frame(x = c("One", "Two", "Three"), y = c(4, 2, 9))
p <- ggplot(df, aes(x = x, y = y, fill = x)) + geom_col()
p + theme_nrc()
}
}
