% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{underemployment_industry}
\alias{underemployment_industry}
\title{Draw employment by industry bar chart, coloured by underemployment levels}
\usage{
underemployment_industry()
}
\value{
ggplot2 object
}
\description{
Draw employment by industry bar chart, coloured by underemployment levels
}
