% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_palettes.R
\name{visualise_palette}
\alias{visualise_palette}
\title{Visualise individual palettes}
\usage{
visualise_palette(palette, n)
}
\arguments{
\item{palette}{Name of colour palette}

\item{n}{Number of colours required}
}
\description{
Visualise individual palettes
}
\examples{
visualise_palette("just orange", n = 5)
}
