% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abs_employment}
\alias{abs_employment}
\title{Employment Time Series}
\usage{
abs_employment(
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Trend"
)
}
\arguments{
\item{states}{which states to include in the plot}

\item{dbf}{how many years does the data span? this adjsuts the x axis scale accordingly}
}
\value{
a ggplot2 object
}
\description{
Employment Time Series
}
