% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plots.R
\name{abs_plot}
\alias{abs_plot}
\title{#This function calls any of the predefined plots in plots.R for convenience}
\usage{
abs_plot(
  indicator,
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_type = "Seasonally Adjusted"
)
}
\arguments{
\item{series_type}{}
}
\value{

}
\description{
#This function calls any of the predefined plots in plots.R for convenience
}
