% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plot.R
\name{abs_plot}
\alias{abs_plot}
\title{Generate a time-series plot for ABS labour force indicators.}
\usage{
abs_plot(
  data = NULL,
  filter_with,
  over = deprecated(),
  years = 2015,
  compare_aus = TRUE,
  markdown = FALSE,
  facet = NULL,
  plotly = FALSE,
  void = FALSE,
  ...
)
}
\arguments{
\item{data}{(optional). Specify a data frame or tibble object to use data other than the labour_force data
included in the \code{aitidata} package. You can use the pipe operator.}

\item{filter_with}{named list specifying what should be plot.}

\item{over}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{over} is no longer supported. Please use \code{filter_with}
instead.}

\item{years}{numeric.}

\item{compare_aus}{(optional) logical. Defaults to TRUE which adds the Australian data for selected indicators.}

\item{markdown}{(optional) logical. Defaults to FALSE. Specify TRUE if you want to use markdown elements in
title/subtitle/axis titles}

\item{facet}{(optional) string. Defaults to NULL. Specify which variable to facet the graph on.}

\item{plotly}{(optional) logical. Defaults to FALSE which creates a ggplot2 plot. Select TRUE to create a plotly plot.
Note that some aspects of the plot are unavailable if plotly = TRUE, including subtitles, and captions.}

\item{void}{(optional) logical. Defaults to FALSE. Specify TRUE to remove all plot elements except for the line.}

\item{...}{other arguments passed to theme_fof()}
}
\value{
A ggplot2 time-series plot or a plotly time-series plot if plotly = TRUE
}
\description{
\code{abs_plot()} can automatically generate appropriate plots for ABS
Time Series indicators for both static display in documents, or RMarkdown,
as well as interactive plots through plotly.
}
