% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_underutilisation.R
\name{abs_underutilisation}
\alias{abs_underutilisation}
\title{Plot Underutilisation levels}
\usage{
abs_underutilisation(
  states,
  years = 5,
  compare_aus = TRUE,
  ages = "Total (age)",
  genders = "Persons",
  series_types = "Seasonally Adjusted"
)
}
\arguments{
\item{states}{Which states/territories to plot. Accepts a character vector if plotting more than one. Plotting
more than one state will index the data to January of the first year included in the data.}

\item{years}{Over how many years the plot should span. Defaults to 5.}

\item{compare_aus}{Whether to add Australian underutilisation to the plot. Default TRUE except if states = "Australia"}

\item{ages}{What age group should be plot. This option is only available if states = "Australia". Defaults to all ages}

\item{genders}{What genders should be plot. Defaults to Persons}

\item{series_types}{Which series type should be plot. Defaults to Seasonally Adjusted}
}
\value{
a ggplot2 object
}
\description{
Plot Underutilisation levels
}
\examples{
The last 5 years hours worked in South Australia: abs_hoursworked("South Australia")
}
