% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fof_palettes.R
\name{aiti_pal}
\alias{aiti_pal}
\title{Generate AITI colours from a palette}
\usage{
aiti_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of colour palette}

\item{reverse}{TRUE to reverse the order of the colours}

\item{...}{Other arguments passed to \code{colorRampPalette()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{aiti_pal()} has been renamed \code{fof_pal()}
}
