% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_percent.R
\name{as_percent}
\alias{as_percent}
\title{Title}
\usage{
as_percent(string, scale = 1, ...)
}
\arguments{
\item{string}{Value to print}

\item{scale}{Scale the value by a multiple}

\item{...}{Other arguments passed to formatC}
}
\value{
character
}
\description{
Title
}
\examples{
as_percent(50) = "50\%"
}
