% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format.R
\name{as_percent}
\alias{as_percent}
\title{Format a value as a pretty string, with a percentage sign printed}
\usage{
as_percent(string, scale = 1, ...)
}
\arguments{
\item{string}{Value to print}

\item{scale}{A value to multiply the number before converting to a string. Default of 1 assumes the value
to be printed has already been multipled by 100.}

\item{...}{Other arguments passed to formatC}
}
\value{
character
}
\description{
Format a value as a pretty string, with a percentage sign printed
}
\examples{
as_percent(50)
as_percent(0.5, scale = 100)

}
