% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{average_over}
\alias{average_over}
\title{The average of an indicator between two years}
\usage{
average_over(data, filter_with, between)
}
\arguments{
\item{data}{a data frame of cleaned ABS Time Series data returned from readabs}

\item{filter_with}{a list of variables to filter the data frame on. Must include an indicator}

\item{between}{a date range \code{c(min, max)} to calculate the average over}
}
\value{
numeric
}
\description{
The average of an indicator between two years
}
\examples{
library(reportabs)
average_over(labour_force, list(indicator = "Employed total"), between = c(2010,2015))

}
