% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{change}
\alias{change}
\title{Report whether an indicator has increased or decreased over the year, or over the previous month}
\usage{
change(data = .data, filter_with = filter_list, type = "id", ym = "year")
}
\arguments{
\item{ym}{}
}
\value{
string
}
\description{
Report whether an indicator has increased or decreased over the year, or over the previous month
}
