% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{current}
\alias{current}
\title{Return the most recent (current) value for an indicator}
\usage{
current(data, filter_with, print = TRUE)
}
\arguments{
\item{data}{a dataframe of cleaned ABS Time Series data returned from readabs}

\item{filter_with}{a list of variables to filter the dataframe on. Valid variables include
gender, age, indicator, and series type.}

\item{print}{logical. If TRUE (default) current() returns a string. Specify FALSE to return a number.}
}
\value{
String if print == TRUE. Number if print == FALSE
}
\description{
Return the most recent (current) value for an indicator
}
\examples{
library(reportabs)
labour_force_briefing <- read_absdata("labour_force_briefing")
current(labour_force_briefing, list(indicator = "Employed total",
series_type = "Seasonally Adjusted"))

}
