% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{employment_growth}
\alias{employment_growth}
\title{Convenience function to draw a bar chart of employment growth, by different types}
\usage{
employment_growth(states, years = 2010)
}
\arguments{
\item{year_since}{}
}
\value{

}
\description{
Convenience function to draw a bar chart of employment growth, by different types
}
