% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labour_force}
\alias{labour_force}
\title{Labour Force}
\format{
A data frame:
\describe{
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{gender}{Gender}
\item{age}{Age bracket}
\item{state}{State}
\item{series_type}{Series Type}
\item{indicator}{Labour force indicator}
\item{value}{Value of indicator}
}
}
\source{
\url{www.abs.gov.au}
}
\usage{
labour_force
}
\description{
Labour force data from ABS Labour Force Survey release 6202.0
Includes data from tables 12, 19, and 23
}
\keyword{datasets}
