% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{last}
\alias{last}
\title{Find the value for an indicator for this month last year, or the previous month.}
\usage{
last(data = .data, filter_with = NULL, ym = "year")
}
\arguments{
\item{filter_with}{}
}
\value{

}
\description{
Find the value for an indicator for this month last year, or the previous month.
}
\examples{

}
