% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R, R/reports.R
\name{last}
\alias{last}
\title{Reporting functions}
\usage{
last(data = ., filter_with = NULL, ym = "year")

last(data = ., filter_with = NULL, ym = "year")
}
\arguments{
\item{filter_with}{}
}
\value{



}
\description{
Reporting functions

Reporting functions
}
\examples{


}
