% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{last_value}
\alias{last_value}
\title{Find the value for an indicator for this month last year, or the previous month.}
\usage{
last_value(data = .data, filter_with = filter_list, ym = "year", print = TRUE)
}
\arguments{
\item{filter_with}{}
}
\value{

}
\description{
Find the value for an indicator for this month last year, or the previous month.
}
\examples{

}
