% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_absdata.R
\name{read_absdata}
\alias{read_absdata}
\title{Read ABS time series data.}
\usage{
read_absdata(name = NULL, export_dir = tempdir())
}
\arguments{
\item{name}{The name, as a string, of the dataset to download.A full list of available data is available at https://github.com/f-fof/aitidata}

\item{export_dir}{The directory in which to save downloaded data. Defaults to a temporary directory.}
}
\value{
data
}
\description{
Read ABS time series data from the \href{https://github.com/f-fof/aitidata}{aitidata} package.
}
