% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_briefing.R
\name{render_monthly_briefing}
\alias{render_monthly_briefing}
\title{Render AITI Monthly Labour Force Briefing in .pdf format}
\usage{
render_monthly_briefing(
  input = system.file("quarto", "monthly_briefing.qmd", package = "reportabs"),
  state = "South Australia",
  output_format
)
}
\arguments{
\item{input}{Path to the RMarkdown file which generates the .pdf.
Defaults to the monthly_briefing file included in the reportabs package.}

\item{state}{String. Name of state (in full). Defaults to South Australia.T}

\item{out_dir}{Directory to save the rendered documents}
}
\value{
One .pdf document
}
\description{
Render AITI Monthly Labour Force Briefing in .pdf format
}
