% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_briefing.R
\name{render_monthly_briefing}
\alias{render_monthly_briefing}
\title{Render AITI Monthly Labour Force Briefing in .pdf format}
\usage{
render_monthly_briefing(
  input = system.file("markdown", "monthly_briefing.Rmd", package = "reportabs"),
  out_dir = "out",
  covid = TRUE,
  state = NULL,
  years = NULL,
  hours_worked = NULL,
  series_type = NULL,
  ...
)
}
\arguments{
\item{input}{Path to the RMarkdown file which generates the .pdf. Defaults to the monthly_briefing file included in the reportabs package}

\item{output_dir}{Directory to output the file}

\item{output_file_name}{Name of output file}
}
\value{
Two .pdf documents
}
\description{
Render AITI Monthly Labour Force Briefing in .pdf format
}
