% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_briefing.R
\name{render_monthly_briefing}
\alias{render_monthly_briefing}
\title{Render AITI Monthly Labour Force Briefing in .pdf format}
\usage{
render_monthly_briefing(
  input = system.file("markdown", "monthly_briefing.Rmd", package = "reportabs"),
  out_dir = "out",
  .hours_worked = TRUE,
  covid = TRUE,
  state = NULL,
  years = NULL,
  series_type = NULL,
  directory = NULL,
  .hours_worked_data = NULL,
  .labour_force_data = NULL
)
}
\arguments{
\item{input}{Path to the RMarkdown file which generates the .pdf. Defaults to the monthly_briefing file included in the reportabs package}

\item{out_dir}{Directory to save the rendered documents}

\item{.hours_worked}{Logical. TRUE (default) to include hours worked data. Defaults to FALSE if state is Northern Territory or Australian Capital Territory}

\item{covid}{Logical. TRUE (Default) for ABS data releases with no Trend series}

\item{state}{String. Name of state (in full). Defaults to South Australia.T}

\item{years}{Number. Included graphs are drawn from January of the year specified. Minimum 1978, and defaults to 2017.}

\item{series_type}{String. Seasonally Adjusted}

\item{directory}{dir}

\item{file}{file}
}
\value{
Two .pdf documents
}
\description{
Render AITI Monthly Labour Force Briefing in .pdf format
}
