% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_briefing.R
\name{render_monthly_briefing}
\alias{render_monthly_briefing}
\title{Render AITI Monthly Labour Force Briefing in .pdf format}
\usage{
render_monthly_briefing(state = "South Australia", output_format)
}
\arguments{
\item{state}{String. Name of state (in full). Defaults to South Australia.}

\item{output_format}{Format to render the document. One of "html" or "pdf"}
}
\description{
Render AITI Monthly Labour Force Briefing in .pdf format
}
