% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahm.R
\name{sahm}
\alias{sahm}
\title{Plot the Sahm Recession Indicator for a given state. The Sahm indicator compares the 3 month
average (seasonally adjusted) unemployment rate with the minimum over the last 12 months.
Sahm values over 1 are indicative
of a recession.}
\usage{
sahm(data, region = "Australia")
}
\arguments{
\item{data}{A dataframe with at least the columns date, state, and value}

\item{region}{The state or territory to plot the indicator for. Accepts Australia}
}
\value{
a ggplot2 object
}
\description{
Plot the Sahm Recession Indicator for a given state. The Sahm indicator compares the 3 month
average (seasonally adjusted) unemployment rate with the minimum over the last 12 months.
Sahm values over 1 are indicative
of a recession.
}
\examples{
\dontrun{sahm("Australia")}

}
