% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_scale.R
\name{scale_colour_fof}
\alias{scale_colour_fof}
\title{Colour scale for Factory of the Future colours}
\usage{
scale_colour_fof(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Name of the colour palette (default "main")}

\item{discrete}{Whether the colours should be discrete or continuous (default TRUE)}

\item{reverse}{Whether the order of the colours should be reversed (default FALSE)}
}
\value{
ggplot2 colour scale
}
\description{
Colour scale for Factory of the Future colours
}
