% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_aiti}
\alias{theme_aiti}
\title{AITI Theme}
\usage{
theme_aiti(
  base_size = 12,
  colour = "Soft Black",
  legend = "none",
  markdown = FALSE,
  flipped = FALSE,
  legacy = FALSE
)
}
\arguments{
\item{base_size}{The base size of text elements of the plot. (default 12)}

\item{colour}{The background colour of the plot.}

\item{legend}{The position of the legend (default no legend)}

\item{markdown}{Whether to use markdown formatting for plot titles (default FALSE)}

\item{flipped}{TRUE to flip the y-axis guide lines to show on the x-axis instead.}

\item{legacy}{Whether to use legacy fonts}
}
\value{
ggplot2 theme
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{theme_aiti()} has been renamed \code{theme_fof()}
}
\details{
Theme for plots in AITI publications, reports, Shiny Apps.
}
