% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reporting.R
\name{value_at}
\alias{value_at}
\title{Value at a specific year/month}
\usage{
value_at(data = .data, filter_with = NULL, month_adjust = 0, at_year, at_month)
}
\arguments{
\item{data}{a dataframe}

\item{filter_with}{a list of variables to filter the dataframe on}

\item{month_adjust}{}

\item{at_year}{a year (numeric)}

\item{at_month}{a month(name)}
}
\value{

}
\description{
Value at a specific year/month
}
